+++ src/main/java/com/mlbbroadcast/team/TeamMaster.java
@@ -23,11 +23,11 @@ public class TeamMaster {
     private Division division;
 
     @Column(name = "external_id", nullable = false)
-    private Long external_id;
+    private int external_id;
 
 
     @Builder
-    public TeamMaster(String name, Division division, Long external_id){
+    public TeamMaster(String name, Division division, int external_id){
         this.name = name;
         this.division = division;
         this.external_id = external_id;
